Fixed #16 - Incorrect test duration with freezegun
freezegun is a Python library used to set the date/time returned by
the Python date/time functions to make testing time-sensitive code
easier.

time.time() is used here to measure the duration of test cases, if
freezegun was in use then the time returned by this could be the
time the test had set up, rather than the actual time. This lead to
test durations being reported as negative (if frozen time is in the
past), or much longer than expected (if fronzen time is in the
future).

freezegun exposes the unmodified time.time() via an alias. This
changes TimingSuite so that it uses this alias if freezegun is
installed.